Move `Operator` trait to a new module
Move the Operator trait from `rten::ops` to `rten::operator`. This removes the
dependency of `rten::graph` on `rten::ops`, potentially enabling the "framework"
code in `rten::graph` to be moved to a separate crate in future.
